fix and update
修复docker-compose run --rm web python manage.py collectstatic --noinput可能出现mysql启动缓慢，导致web连接数据库失败问题，优化docker-compose配置

+++ configs/default/sites.py
@@ -27,17 +27,17 @@ DEFAULT_THEME = os.getenv("DEFAULT_THEME", "default")
 DEFAULT_METHOD = os.getenv("DEFAULT_METHOD", "aes-256-cfb")
 
 # 签到流量设置
-MIN_CHECKIN_TRAFFIC = os.getenv("MIN_CHECKIN_TRAFFIC", 10 * MB)
-MAX_CHECKIN_TRAFFIC = os.getenv("MAX_CHECKIN_TRAFFIC", 200 * MB)
+MIN_CHECKIN_TRAFFIC = int(os.getenv("MIN_CHECKIN_TRAFFIC", 10 * MB))
+MAX_CHECKIN_TRAFFIC = int(os.getenv("MAX_CHECKIN_TRAFFIC", 200 * MB))
 
 # 网站title
 TITLE = os.getenv("TITLE", "谜之屋")
 SUBTITLE = os.getenv("SUBTITLE", "秘密的小屋")
 
 # 用户邀请返利比例
-INVITE_PERCENT = os.getenv("INVITE_PERCENT", 0.2)
+INVITE_PERCENT = float(os.getenv("INVITE_PERCENT", 0.2))
 # 用户能生成的邀请码数量
-INVITE_NUM = os.getenv("INVITE_NUM ", 5)
+INVITE_NUM = int(os.getenv("INVITE_NUM ", 5))
 
 # 网站邀请界面提示语
 INVITEINFO = os.getenv("INVITEINFO", "邀请码实时更新，如果用完了就没了")
